requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
flask-mail
flask-wtf
flask-talisman
flask-limiter
python-dotenv
werkzeug
reportlab
gunicorn
psycopg2-binary
